Use more targeted NetworkPolicies (#25)
In #4, we added a NetworkPolicy. The intent was to prevent the Redis
and/or Sentinel pods from differing RedisFailovers from joining up with
one another (See:
https://github.com/spotahome/redis-operator/issues/550). These policies
have proven to be too coarsely grained. We end up deploying supplemental
NetworkPolicies to allow ingress traffic.

This change narrows the scope of the NetworkPolicies manged by the
operator. One policy allows traffic to the Redis node pods ONLY on the
redis port and monitoring port for traffic originating from within the
namespace. The other policy allows traffic to the Sentinel pods ONLY on
the sentinel port for traffic originating from within the namespace. All
other traffic to these pods will be dropped.

Connections to the HAProxy pods - IE access to the redis master node -
will now be allowed by default. This achieves the goals of #4 and allows
us to stop littering additional NetworkPolicies to allow external
communication with a Redis instance.