Fix resource name generation
Use the operator's canonical name generation pattern and faculties
when generating haproxy resource names. This allows us to remove a
parallel, bespoke name generator from the operator source.

The haproxy resources were using the name pattern:

```
redis-haproxy-[redis-failover-name]
```

The operator's reource name generator tries to add as little as
possible to resource names to avoid running into the character limits
imposed by Kubernetes. We should strive to do the same with the
resources we add to the operator. With this change, the operator
generates haproxy resources with the following naming pattern
indicating that the resource relates to the redisfailvoer redis'
proxy (rfrp):

```
rfrp-[redis-failover-name]
```

The headless service is also promoted to a first class concern of the
name generator to avoid collisions with other services placed by the
operator. It is now named using the following pattern:

```
rfrh-[redis-failvoer-name]
```