Remove orphaned network policy resource (#49)
Fixes: https://github.com/powerhome/redis-operator/issues/48

This PR aims to address an issue introduced in release 2.0.1, where the
logic for generating network policy resources by the operator was
changed. From version 2.0.1 onwards, the operator no longer generates
network policies for Redis but failed to remove existing policies
generated prior to this version. This PR introduces a check and removal
process for that leftover network policiy.

---------

Co-authored-by: Ben Langfeld <ben@langfeld.me>
Co-authored-by: Aaron Kuehler <aaron.kuehler@powerhrg.com>