Prevent connections to replica on replica restart (#57)
There is a brief period of time during a Redis replica restart when
HAProxy can send writes to it; causing the following error:

```
READONLY You can't write against a read only replica.
```

This is because HAProxy (< 3.1.0) always treats newly detected backends
as immediately UP - ready to serve traffic - UNTIL they failed their
checks. HAProxy v3.1.0 adds the ability to configure a backend's initial
state - init-state - and how/when it HAProxy determines that the backend
is ready to receive traffic.

This changes the HAProxy configuration so that a new Redis node is
disqualified from client receiving traffic until HAProxy determines that
the node is indeed a "master" node.

This also updates the default operator's HAProxy image accordingly to
receive the `init-state` HAProxy configuration option.

References:

- https://github.com/haproxy/haproxy/issues/51
-
https://github.com/haproxy/haproxy/commit/50322dff81f3c05e67bf311ff9ef0c77c833df97