haproxy: disconnect clients when node is demoted to replica (#59)
Some clients do not gracefully handle when a failover occurs; they keep
the established connection open. When the redis node comes back up as a
replica, the client recieves an error indicating that they cannot write
against the replica:

```
READONLY You can't write against a read only replica
```

This closes client connections to a redis node when haproxy detects that
it is in a "down" state. The idea is that the clients will recognize
they need to reestablish a connection.

References:
- https://docs.haproxy.org/3.1/configuration.html#5.2-on-marked-down